Renaming models directory for the elastic agent configurations
