+++ server/webapp/WEB-INF/rails/webpack/models/elastic_agent_configurations/cluster_profiles_crud.ts
