+++ server/webapp/WEB-INF/rails/webpack/models/elastic_agent_configurations/elastic_agent_profiles_crud.ts
