+++ server/webapp/WEB-INF/rails/webpack/models/elastic_agent_configurations/types.ts
