+++ server/webapp/WEB-INF/rails/webpack/views/pages/elastic_agent_configurations.tsx
@@ -17,9 +17,9 @@ import {ErrorResponse} from "helpers/api_request_builder";
 import * as m from "mithril";
 import {Stream} from "mithril/stream";
 import * as stream from "mithril/stream";
-import {ClusterProfilesCRUD} from "models/elastic_profiles/cluster_profiles_crud";
-import {ElasticAgentProfilesCRUD} from "models/elastic_profiles/elastic_agent_profiles_crud";
-import {ClusterProfile, ClusterProfiles, ElasticAgentProfile, ElasticAgentProfiles} from "models/elastic_profiles/types";
+import {ClusterProfilesCRUD} from "models/elastic_agent_configurations/cluster_profiles_crud";
+import {ElasticAgentProfilesCRUD} from "models/elastic_agent_configurations/elastic_agent_profiles_crud";
+import {ClusterProfile, ClusterProfiles, ElasticAgentProfile, ElasticAgentProfiles} from "models/elastic_agent_configurations/types";
 import {ExtensionType} from "models/shared/plugin_infos_new/extension_type";
 import {Extension} from "models/shared/plugin_infos_new/extensions";
 import {PluginInfo} from "models/shared/plugin_infos_new/plugin_info";
