+++ server/webapp/WEB-INF/rails/webpack/views/pages/elastic_agent_configurations/cluster_profiles_modals.tsx
@@ -21,8 +21,8 @@ import * as _ from "lodash";
 import * as m from "mithril";
 import {Stream} from "mithril/stream";
 import * as stream from "mithril/stream";
-import {ClusterProfilesCRUD} from "models/elastic_profiles/cluster_profiles_crud";
-import {ClusterProfile} from "models/elastic_profiles/types";
+import {ClusterProfilesCRUD} from "models/elastic_agent_configurations/cluster_profiles_crud";
+import {ClusterProfile} from "models/elastic_agent_configurations/types";
 import {Configurations} from "models/shared/configuration";
 import {ExtensionType} from "models/shared/plugin_infos_new/extension_type";
 import {ElasticAgentSettings, Extension} from "models/shared/plugin_infos_new/extensions";
