+++ server/webapp/WEB-INF/rails/webpack/views/pages/elastic_agent_configurations/spec/elastic_profile_widget_spec.tsx
@@ -15,7 +15,7 @@
  */
 import * as _ from "lodash";
 import * as m from "mithril";
-import {ElasticAgentProfile} from "models/elastic_profiles/types";
+import {ElasticAgentProfile} from "models/elastic_agent_configurations/types";
 import {Extension} from "models/shared/plugin_infos_new/extensions";
 import {PluginInfo} from "models/shared/plugin_infos_new/plugin_info";
